# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Device Uniqueness Backend",
    "description": "**Prove the device and register the username on People Chain.**\n\n### Validation flow\n\nThe iOS smoke drives this sequence end-to-end; each step gates the next:\n\n1. `POST /api/v1/auth/challenges` — mint a single-use challenge\n2. `POST /api/v1/auth/app-attest/attestations` — iOS App Attest (`202`, no-op in M0)\n3. `POST /api/v1/auth/token` — sr25519 proof exchanged for a JWT\n4. `POST /api/v1/usernames/available` — JWT-gated People Chain read\n5. `POST /api/v1/usernames` — `202` outbox reservation\n\n> **Attestation is disabled for M0.** Apple/Google verification is intentionally a no-op; the sr25519 account proof and People Chain reads still run end-to-end.\n\nErrors come in two shapes. The auth handshake (`/api/v1/auth/*`) answers the `ErrorResponse` envelope — `{ error, message }`, where `error` is a stable machine code. Every other route answers `{ error }`, a human-readable summary, plus a `fields` array of `{ field, message }` when the failure is per-field validation.",
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "2"
  },
  "servers": [
    {
      "url": "https://identity.dotspark.app",
      "description": "Test deployment"
    }
  ],
  "paths": {
...
    "/api/v1/turn/issue": {
      "post": {
        "tags": [
          "TURN"
        ],
        "summary": "Issue short-lived TURN credentials for WebRTC ICE negotiation (JWT-gated).",
        "operationId": "issue_credentials",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Credentials minted: `username` is `{unixExpiry}:{hexId}` (expiry = now + TTL), `password` is the base64 HMAC over `username` under the secret shared with the TURN relay (the coturn REST-API construction), `servers` echoes the configured ICE server list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Body validation failed (with per-field `fields`), or `Malformed JSON in request body` when a body is present but not JSON. An empty body is accepted — every field is optional."
          },
          "401": {
            "description": "Missing/malformed `Authorization` header or failed token verification."
          },
          "429": {
            "description": "Per-subject rate limit exceeded (with `Retry-After`)."
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      }
    },
    "/api/v1/turn/issue-with-proof": {
      "post": {
        "tags": [
          "TURN"
        ],
        "summary": "Redeem a personhood proof for the same credentials as `/turn/issue`.",
        "operationId": "issue_with_proof",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueWithProofBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Proof accepted; returns servers, username, password, and the configured TTL. Credentials expire `ttl` seconds after issuance; no alias appears in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unparseable body, invalid hex, a collection outside the canonical People Lite/People allowlist, a proof that is not a single-context ring-VRF signature, a product this deployment does not accept, or a timestamp outside the accepted skew."
          },
          "403": {
            "description": "The proof did not verify against the named ring root under that product's context — including when `(ringIndex, ringRevision)` is outside the roots this deployment still holds (deliberately unspecific)."
          },
          "429": {
            "description": "This person's redemption limit was exceeded (with `Retry-After`)."
          },
          "503": {
            "description": "Verification unavailable: no ring-root snapshot yet (chain unreachable since boot), the bounded waiter queue is full, or all verification slots remained busy for the bounded wait. Saturation responses include `Retry-After`."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IssueRequest": {
        "type": "object",
        "description": "`{ regionHint? }` — the issue command (the hint is reserved and ignored).",
        "properties": {
          "regionHint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional region hint (reserved for future use; accepted and ignored).",
            "example": "eu-west"
          }
        }
      },
      "IssueResponse": {
        "type": "object",
        "description": "The 201 body: coturn REST-API ephemeral credentials plus the ICE\nserver list.",
        "required": [
          "servers",
          "username",
          "password",
          "ttl"
        ],
        "properties": {
          "servers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Configured ICE server URLs (`stun:` / `turn:` forms), echoed verbatim.",
            "example": [
              "stun:stun.example.com:3478",
              "turn:turn.example.com:3478?transport=udp"
            ]
          },
          "username": {
            "type": "string",
            "description": "`{unixExpiry}:{hexId}`. JWT issuance uses mint time + configured TTL\nand 8 random id bytes; proof issuance uses mint time + configured TTL\nand a deterministic opaque 16-byte id derived from product and alias.",
            "example": "1784757652:0a79e3412921701a"
          },
          "password": {
            "type": "string",
            "description": "Base64 HMAC over `username` under the relay-shared secret (algorithm\nper deployment config, default HMAC-SHA1).",
            "example": "qmg5g7d1bXzY0qZkRUqtIPEIKjA="
          },
          "ttl": {
            "type": "integer",
            "format": "int64",
            "description": "Credential time-to-live in whole seconds (the configured `ttl_secs`).",
            "example": 1800,
            "minimum": 0
          }
        }
      },
      "IssueWithProofBody": {
        "type": "object",
        "description": "Proof-redemption request.",
        "required": [
          "productId",
          "collection",
          "proof",
          "ringIndex",
          "ringRevision",
          "timestamp"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "description": "The product the proof was made for. Its context is the one the proof\nis verified under, so a product proves under its own identifier. Must\nbe one this deployment accepts.",
            "example": "vox.dot"
          },
          "collection": {
            "type": "string",
            "description": "Hex-encoded 32-byte collection id from the proof's TrUAPI\n`ringLocation`. Only the canonical People Lite and People collections\nare accepted.",
            "example": "0x706f703a706f6c6b61646f742e6e6574776f726b2f70656f706c652d6c697465"
          },
          "proof": {
            "type": "string",
            "description": "Ring-VRF proof over the derived message (hex), exactly as the host\nreturns it — raw signature bytes, with no SCALE length prefix."
          },
          "ringIndex": {
            "type": "integer",
            "format": "int32",
            "description": "Ring containing the proving member, used to locate the root.",
            "example": 0,
            "minimum": 0
          },
          "ringRevision": {
            "type": "integer",
            "format": "int32",
            "description": "Revision of that ring the proof was made against, as the host reports\nit. Only that revision's root is tried, so a revision this deployment\nno longer holds is refused without verifying anything.",
            "example": 3,
            "minimum": 0
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Client Unix seconds, bound into the proved message. Must be within the\nserver's accepted skew.",
            "example": 1784757652,
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access JWT issued by POST /api/v1/auth/token."
      }
    }
  },
...
}